Updating ItemExtensions.cs to better handle Linkfields
This change was in response to some seemingly random errors we were getting when on a project when trying to use the LinkFieldUrl() function. It was occasionally throwing errors complaining about a missing "url" parameter. This seemed to be because older, XML based link generation functions were being used. This commit updates this function to be more robust and properly handle additional linkfield options (media, javascript, etc...) as well.